(doc, refactor): add comments, rename assignments methods
