+++ src/main/scala/satify/model/dpll/PartialAssignment.scala
