+++ src/main/scala/satify/update/solver/dpll/DpllDecision.scala
@@ -6,7 +6,7 @@ import satify.model.cnf.CNF.Symbol
 import satify.model.dpll.{Constraint, Decision}
 import satify.update.solver.dpll.Optimizations.{pureLiteralIdentification, unitLiteralIdentification}
 import satify.update.solver.dpll.cnf.CNFSimplification.simplifyCnf
-import satify.update.solver.dpll.utils.PartialModelUtils.{filterUnconstrVars, updatePartialAssignment}
+import satify.update.solver.dpll.utils.PartialAssignmentUtils.{filterUnconstrVars, updatePartialAssignment}
 
 import scala.util.Random
